pyproject.toml
[tool.poetry]
name = "weetags"
version = "0.1.0"
description = ""
authors = ["Romain Viry <rom88.viry@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sanic = "^23.12.1"
treelib = "1.6.4"
pyjwt = "^2.7.0"
pyyaml = "^6.0"
sanic-ext = "^23.12.0"
attrs = "^23.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
